Use the RSpec 1 syntax with `should`
The test suite use the RSpec 1 syntax with `should` that got deprecated.
In order to ease up the transition, for now we want to stick with the
`should` syntax, and *not* use `expect`.

While here, also ensure exceptions are raised when deprecated features
are used and fix a bunch of warnings.